Rename channel files to enable proper autoloading
The autoloading mechanism was expecting to define a constant:
`MatestackUiVuejsChannel` given the file name `matestack_ui_vuejs_channel`.
When the subscriber was attempting to subscribe to `MatestackUiVueJsChannel`, it
was unable to find said constant under app/channels.

For parity, the javascript/channels subscriber counterpart was also renamed
as well as the stream name.